backend: streamline functions to actual DB values
Parameters in stored procedures were outdated regarding the values present in the DB.
Align these to a 1:1 equality.

+++ packaging/dbscripts/clusters_sp.sql
@@ -582,7 +582,7 @@ LANGUAGE plpgsql;
 --Returns true if update was successful.
 CREATE OR REPLACE FUNCTION UpdateClusterUpgradeProgress (
     v_cluster_id UUID,
-    v_upgrade_percent_complete INT,
+    v_upgrade_percent_complete SMALLINT,
     OUT v_updated BOOLEAN
     )
 AS $FUNCTION$
